+++ pdata/internal/generated_pmetric_test.go
@@ -931,14 +931,6 @@ func TestNumberDataPoint_Exemplars(t *testing.T) {
 	assert.EqualValues(t, testValExemplars, ms.Exemplars())
 }
 
-func TestNumberDataPoint_Flags(t *testing.T) {
-	ms := NewNumberDataPoint()
-	assert.EqualValues(t, MetricDataPointFlagsNone, ms.Flags())
-	testValFlags := MetricDataPointFlagsNone
-	ms.SetFlags(testValFlags)
-	assert.EqualValues(t, testValFlags, ms.Flags())
-}
-
 func TestNumberDataPoint_FlagsStruct(t *testing.T) {
 	ms := NewNumberDataPoint()
 	fillTestMetricDataPointFlagsStruct(ms.FlagsStruct())
@@ -1135,14 +1127,6 @@ func TestHistogramDataPoint_Exemplars(t *testing.T) {
 	fillTestExemplarSlice(ms.Exemplars())
 	testValExemplars := generateTestExemplarSlice()
 	assert.EqualValues(t, testValExemplars, ms.Exemplars())
-}
-
-func TestHistogramDataPoint_Flags(t *testing.T) {
-	ms := NewHistogramDataPoint()
-	assert.EqualValues(t, MetricDataPointFlagsNone, ms.Flags())
-	testValFlags := MetricDataPointFlagsNone
-	ms.SetFlags(testValFlags)
-	assert.EqualValues(t, testValFlags, ms.Flags())
 }
 
 func TestHistogramDataPoint_FlagsStruct(t *testing.T) {
@@ -1369,14 +1353,6 @@ func TestExponentialHistogramDataPoint_Exemplars(t *testing.T) {
 	fillTestExemplarSlice(ms.Exemplars())
 	testValExemplars := generateTestExemplarSlice()
 	assert.EqualValues(t, testValExemplars, ms.Exemplars())
-}
-
-func TestExponentialHistogramDataPoint_Flags(t *testing.T) {
-	ms := NewExponentialHistogramDataPoint()
-	assert.EqualValues(t, MetricDataPointFlagsNone, ms.Flags())
-	testValFlags := MetricDataPointFlagsNone
-	ms.SetFlags(testValFlags)
-	assert.EqualValues(t, testValFlags, ms.Flags())
 }
 
 func TestExponentialHistogramDataPoint_FlagsStruct(t *testing.T) {
@@ -1609,14 +1585,6 @@ func TestSummaryDataPoint_QuantileValues(t *testing.T) {
 	fillTestValueAtQuantileSlice(ms.QuantileValues())
 	testValQuantileValues := generateTestValueAtQuantileSlice()
 	assert.EqualValues(t, testValQuantileValues, ms.QuantileValues())
-}
-
-func TestSummaryDataPoint_Flags(t *testing.T) {
-	ms := NewSummaryDataPoint()
-	assert.EqualValues(t, MetricDataPointFlagsNone, ms.Flags())
-	testValFlags := MetricDataPointFlagsNone
-	ms.SetFlags(testValFlags)
-	assert.EqualValues(t, testValFlags, ms.Flags())
 }
 
 func TestSummaryDataPoint_FlagsStruct(t *testing.T) {
@@ -2106,7 +2074,6 @@ func fillTestNumberDataPoint(tv NumberDataPoint) {
 	tv.SetTimestamp(Timestamp(1234567890))
 	tv.SetDoubleVal(float64(17.13))
 	fillTestExemplarSlice(tv.Exemplars())
-	tv.SetFlags(MetricDataPointFlagsNone)
 	fillTestMetricDataPointFlagsStruct(tv.FlagsStruct())
 }
 
@@ -2139,7 +2106,6 @@ func fillTestHistogramDataPoint(tv HistogramDataPoint) {
 	tv.SetBucketCounts(NewImmutableUInt64Slice([]uint64{1, 2, 3}))
 	tv.SetExplicitBounds(NewImmutableFloat64Slice([]float64{1, 2, 3}))
 	fillTestExemplarSlice(tv.Exemplars())
-	tv.SetFlags(MetricDataPointFlagsNone)
 	fillTestMetricDataPointFlagsStruct(tv.FlagsStruct())
 	tv.SetMin(float64(9.23))
 	tv.SetMax(float64(182.55))
@@ -2176,7 +2142,6 @@ func fillTestExponentialHistogramDataPoint(tv ExponentialHistogramDataPoint) {
 	fillTestBuckets(tv.Positive())
 	fillTestBuckets(tv.Negative())
 	fillTestExemplarSlice(tv.Exemplars())
-	tv.SetFlags(MetricDataPointFlagsNone)
 	fillTestMetricDataPointFlagsStruct(tv.FlagsStruct())
 	tv.SetMin(float64(9.23))
 	tv.SetMax(float64(182.55))
@@ -2220,7 +2185,6 @@ func fillTestSummaryDataPoint(tv SummaryDataPoint) {
 	tv.SetCount(uint64(17))
 	tv.SetSum(float64(17.13))
 	fillTestValueAtQuantileSlice(tv.QuantileValues())
-	tv.SetFlags(MetricDataPointFlagsNone)
 	fillTestMetricDataPointFlagsStruct(tv.FlagsStruct())
 }