+++ pdata/internal/metrics.go
@@ -183,6 +183,54 @@ func (at MetricAggregationTemporality) String() string {
 	return otlpmetrics.AggregationTemporality(at).String()
 }
 
+// Flags returns the flags associated with this NumberDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms NumberDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.orig.Flags)
+}
+
+// SetFlags replaces the flags associated with this NumberDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms NumberDataPoint) SetFlags(v MetricDataPointFlags) {
+	ms.orig.Flags = uint32(v)
+}
+
+// Flags returns the flags associated with this HistogramDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms HistogramDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.orig.Flags)
+}
+
+// SetFlags replaces the flags associated with this HistogramDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms HistogramDataPoint) SetFlags(v MetricDataPointFlags) {
+	ms.orig.Flags = uint32(v)
+}
+
+// Flags returns the flags associated with this ExponentialHistogramDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms ExponentialHistogramDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.orig.Flags)
+}
+
+// SetFlags replaces the flags associated with this ExponentialHistogramDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms ExponentialHistogramDataPoint) SetFlags(v MetricDataPointFlags) {
+	ms.orig.Flags = uint32(v)
+}
+
+// Flags returns the flags associated with this SummaryDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms SummaryDataPoint) Flags() MetricDataPointFlags {
+	return MetricDataPointFlags(ms.orig.Flags)
+}
+
+// SetFlags replaces the flags associated with this SummaryDataPoint.
+// Deprecated: [v0.57.0] Use FlagsStruct instead.
+func (ms SummaryDataPoint) SetFlags(v MetricDataPointFlags) {
+	ms.orig.Flags = uint32(v)
+}
+
 // MetricDataPointFlagsStruct defines how a metric aggregator reports aggregated values.
 // It describes how those values relate to the time interval over which they are aggregated.
 //